refactor(Table): aggregate 단위로 패키지 묶기
menu, menuGroup 패키지 통합
