+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -3,12 +3,11 @@ package kitchenpos.menu.application;
 import kitchenpos.exception.MenuException;
 import kitchenpos.exception.ProductException;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProductRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Products;
 import org.springframework.stereotype.Service;
