+++ src/main/java/kitchenpos/menu/domain/MenuGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.menugroup.domain;
+package kitchenpos.menu.domain;
 
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
