+++ src/test/java/kitchenpos/menu/domain/MenuGroupTest.java
@@ -1,5 +1,6 @@
-package kitchenpos.menugroup.domain;
+package kitchenpos.menu.domain;
 
+import kitchenpos.menu.domain.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
