+++ src/test/java/kitchenpos/menu/dto/MenuGroupResponseTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.menugroup.dto;
+package kitchenpos.menu.dto;
 
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.dto.MenuGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
